+++ app/src/main/kotlin/no/nav/k9punsj/K9PunsjApplication.kt
@@ -6,9 +6,7 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import de.huxhorn.sulky.ulid.ULID
 import no.nav.k9punsj.db.config.DbConfiguration
 import no.nav.k9punsj.db.config.hikariConfig
-import no.nav.k9punsj.db.config.hikariConfigLocal
 import no.nav.k9punsj.jackson.UlidDeserializer
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.Banner
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.autoconfigure.flyway.FlywayAutoConfiguration
@@ -17,15 +15,13 @@ import org.springframework.boot.runApplication
 import org.springframework.boot.web.embedded.netty.NettyReactiveWebServerFactory
 import org.springframework.boot.web.reactive.server.ReactiveWebServerFactory
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Profile
-import org.springframework.core.env.Environment
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder
 import org.springframework.scheduling.annotation.EnableScheduling
 import javax.sql.DataSource
 
 @EnableScheduling
 @SpringBootApplication(exclude = [ErrorMvcAutoConfiguration::class, FlywayAutoConfiguration::class])
-class K9PunsjApplication @Autowired constructor(var dbConfiguration: DbConfiguration, val environment: Environment) {
+class K9PunsjApplication {
 
 	@Bean
 	fun objectMapperBuilder(): Jackson2ObjectMapperBuilder {
@@ -42,15 +38,9 @@ class K9PunsjApplication @Autowired constructor(var dbConfiguration: DbConfigura
 	}
 
 	@Bean
-	@Profile("!test & !local")
-	fun databaseInitializer(): DataSource {
+	@StandardProfil
+	fun databaseInitializer(dbConfiguration: DbConfiguration): DataSource {
 		return hikariConfig(dbConfiguration)
-	}
-
-	@Bean
-	@Profile("local")
-	fun databaseInitializerLocal(): DataSource {
-		return hikariConfigLocal(dbConfiguration, environment)
 	}
 }
