+++ app/src/main/kotlin/no/nav/k9punsj/StandardProfil.kt
@@ -1,0 +1,6 @@
+package no.nav.k9punsj
+
+import org.springframework.context.annotation.Profile
+
+@Profile("!test & !local")
+annotation class StandardProfil
