+++ app/src/main/kotlin/no/nav/k9punsj/db/config/HikariConfig.kt
@@ -2,7 +2,6 @@ package no.nav.k9punsj.db.config
 
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
-import org.springframework.core.env.Environment
 
 fun createHikariConfig(
         jdbcUrl: String,
@@ -23,10 +22,5 @@ fun createHikariConfig(
 
 fun hikariConfig(hikariConfig: DbConfiguration): HikariDataSource {
     migrate(hikariConfig)
-    return getDataSource(hikariConfig)
-}
-
-fun hikariConfigLocal(hikariConfig: DbConfiguration, environment: Environment): HikariDataSource {
-    runMigrationLocal(hikariConfig, environment)
     return getDataSource(hikariConfig)
 }
