+++ app/src/test/kotlin/no/nav/k9punsj/TestProfil.kt
@@ -1,0 +1,6 @@
+package no.nav.k9punsj
+
+import org.springframework.context.annotation.Profile
+
+@Profile("test")
+annotation class TestProfil
